+++ src/constants.py
@@ -1,0 +1,4 @@
+MODEL_SAMPLE_RATE: int = 16000
+CHUNK_SIZE: int = 30
+BATCH_SIZE: int = 4
+CONFIDENCE_THRESHOLD: float = 0.5
